pyproject.toml
[project]
name = "ragdocumentqawithgroqapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "arxiv>=2.4.1",
    "bs4>=0.0.2",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.28",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "langserve>=0.3.3",
    "pdf2image>=1.17.0",
    "pypdf>=6.10.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.0",
    "sse-starlette>=3.3.4",
    "streamlit>=1.56.0",
    "tools>=1.0.26",
    "torchvision>=0.26.0",
    "unstructured>=0.22.18",
    "uvicorn>=0.44.0",
]
